requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.0
django-debug-toolbar==2.2
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
idna==2.9
mysqlclient==1.4.6
Pillow==7.0.0
PyMySQL==0.9.3
python-memcached==1.59
pytz==2019.3
qiniu==7.2.8
requests==2.23.0
shortuuid==1.0.1
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
